Rework design of variable time-order handling
As discussed with @roystgnr, these things really belong in the DiffSystem.
So this moves these things around from the TimeSolver to the DiffSystem.
A pleasant side-effect is that this really cleans up the user-facing
API in the FEMSystem subclasses, see the changes in fem_system_ex3
in this commit.
