+++ include/physics/diff_physics.h
@@ -245,11 +245,10 @@ public:
    * and specify the order-in-time of the variable, either 1 or 2. This method
    * assumes the variable is first order for backward compatibility.
    */
-  virtual void time_evolving (unsigned int var) {
+  virtual void time_evolving (unsigned int var)
+  {
     libmesh_deprecated();
-    if (_time_evolving.size() <= var)
-      _time_evolving.resize(var+1, 0);
-    _time_evolving[var] = 1;
+    this->time_evolving(var,1);
   }
 
   /**
@@ -273,12 +272,13 @@ public:
    * behave like du/dt = F(u), and should not call time_evolving()
    * for any variables which behave like 0 = G(u).
    */
-  bool is_time_evolving (unsigned int var) const {
+  bool is_time_evolving (unsigned int var) const
+  {
     libmesh_assert_less(var,_time_evolving.size());
     libmesh_assert( _time_evolving[var] == 0 ||
                     _time_evolving[var] == 1 ||
                     _time_evolving[var] == 2 );
-    return !(_time_evolving[var] == 0);
+    return _time_evolving[var];
   }
 
   /**
@@ -506,6 +506,31 @@ public:
   bool _eulerian_time_deriv (bool request_jacobian,
                              DiffContext &);
 
+  bool have_first_order_vars() const
+  { return !_first_order_vars.empty(); }
+
+  /**
+   * Returns the set of first order in time variable indices. May be empty.
+   */
+  const std::set<unsigned int> & get_first_order_vars() const
+  { return _first_order_vars; }
+
+  bool is_first_order_var( unsigned int var ) const
+  { return _first_order_vars.find(var) != _first_order_vars.end(); }
+
+
+  bool have_second_order_vars() const
+  { return !_second_order_vars.empty(); }
+
+  /**
+   * Returns the set of second order in time variable indices. May be empty.
+   */
+  const std::set<unsigned int> & get_second_order_vars() const
+  { return _second_order_vars; }
+
+  bool is_second_order_var( unsigned int var ) const
+  { return _second_order_vars.find(var) != _second_order_vars.end(); }
+
 
 protected:
 
@@ -526,9 +551,24 @@ protected:
    */
   std::vector<unsigned int> _time_evolving;
 
+  /**
+   * Variable indices for those variables that are first order in time.
+   */
+  std::set<unsigned int> _first_order_vars;
+
+  /**
+   * Variable indices for those variables that are second order in time.
+   */
+  std::set<unsigned int> _second_order_vars;
+
+  /**
+   * If the user adds any second order variables, then we need to also
+   * cache the map to their corresponding dot variable that will
+   * be added by this TimeSolver class.
+   */
+  std::map<unsigned int,unsigned int> _second_order_dot_vars;
+
 };
-
-
 
 // ------------------------------------------------------------
 // DifferentiablePhysics inline methods
