+++ src/solvers/first_order_unsteady_solver.C
@@ -18,52 +18,9 @@
 #include "libmesh/first_order_unsteady_solver.h"
 #include "libmesh/diff_system.h"
 #include "libmesh/quadrature.h"
-#include "libmesh/dirichlet_boundaries.h"
-#include "libmesh/dof_map.h"
 
 namespace libMesh
 {
-
-void FirstOrderUnsteadySolver::init()
-{
-  // First call parent init
-  UnsteadySolver::init();
-
-  // Now check for second order variables and add their velocities to the System.
-  const std::set<unsigned int> & second_order_vars = this->get_second_order_vars();
-  if( !second_order_vars.empty() )
-    {
-      for( std::set<unsigned int>::const_iterator var_it = second_order_vars.begin();
-           var_it != second_order_vars.end(); ++var_it )
-        {
-          const Variable & var = _system.variable(*var_it);
-          std::string new_var_name = std::string("dot_")+var.name();
-
-          unsigned int v_var_idx;
-
-          if( var.active_subdomains().empty() )
-            v_var_idx = _system.add_variable( new_var_name, var.type() );
-          else
-            v_var_idx = _system.add_variable( new_var_name, var.type(), &var.active_subdomains() );
-
-          _second_order_dot_vars.insert( std::pair<unsigned int,unsigned int>(*var_it,v_var_idx) );
-
-          // The new velocities are time evolving variables of first order
-          this->system().time_evolving( v_var_idx, 1 );
-
-          // And if there are any boundary conditions set on the second order
-          // variable, we also need to set it on its velocity variable.
-          this->add_dot_var_dirichlet_bcs( *var_it, v_var_idx );
-        }
-    }
-}
-
-unsigned int FirstOrderUnsteadySolver::get_second_order_dot_var(unsigned int var) const
-{
-  libmesh_assert(this->is_second_order_var(var));
-  libmesh_assert(_second_order_dot_vars.find(var) != _second_order_dot_vars.end());
-  return _second_order_dot_vars.find(var)->second;
-}
 
 void FirstOrderUnsteadySolver::prepare_accel(DiffContext & context)
 {
@@ -80,10 +37,10 @@ bool FirstOrderUnsteadySolver::compute_second_order_eqns(bool compute_jacobian, 
 
   for (unsigned int var = 0; var != context.n_vars(); ++var)
     {
-      if (!this->is_second_order_var(var))
+      if (!this->_system.is_second_order_var(var))
         continue;
 
-      unsigned int dot_var = this->get_second_order_dot_var(var);
+      unsigned int dot_var = this->_system.get_second_order_dot_var(var);
 
       // We're assuming that the FE space for var and dot_var are the same
       libmesh_assert( context.get_system().variable(var).type() ==
@@ -135,94 +92,6 @@ bool FirstOrderUnsteadySolver::compute_second_order_eqns(bool compute_jacobian, 
     }
 
   return compute_jacobian;
-}
-
-
-void FirstOrderUnsteadySolver::add_dot_var_dirichlet_bcs( unsigned int var_idx,
-                                                          unsigned int dot_var_idx )
-{
-  // We're assuming that there could be a lot more variables than
-  // boundary conditions, so we search each of the boundary conditions
-  // for this variable rather than looping over boundary conditions
-  // in a separate loop and searching through all the variables.
-  const DirichletBoundaries * all_dbcs =
-    this->system().get_dof_map().get_dirichlet_boundaries();
-
-  if( all_dbcs )
-    {
-      // We need to cache the DBCs to be added so that we add them
-      // after looping over the existing DBCs. Otherwise, we're polluting
-      // the thing we're looping over.
-      std::vector<DirichletBoundary*> new_dbcs;
-
-      DirichletBoundaries::const_iterator dbc_it = all_dbcs->begin();
-      for( ; dbc_it != all_dbcs->end(); ++dbc_it )
-        {
-          libmesh_assert(*dbc_it);
-          DirichletBoundary & dbc = *(*dbc_it);
-
-          // Look for second order variable in the current
-          // DirichletBoundary object
-          std::vector<unsigned int>::const_iterator dbc_var_it =
-            std::find( dbc.variables.begin(), dbc.variables.end(), var_idx );
-
-          // If we found it, then we also need to add it's corresponding
-          // "dot" variable to a DirichletBoundary
-          std::vector<unsigned int> vars_to_add;
-          if( dbc_var_it != dbc.variables.end() )
-            vars_to_add.push_back(dot_var_idx);
-
-          if( !vars_to_add.empty() )
-            {
-              DirichletBoundary * new_dbc;
-
-              if( dbc.f )
-                {
-                  if( dbc.g )
-                    new_dbc = new DirichletBoundary(dbc.b,
-                                                    vars_to_add,
-                                                    dbc.f.get(),
-                                                    dbc.g.get());
-
-                  else
-                    new_dbc = new DirichletBoundary(dbc.b,
-                                                    vars_to_add,
-                                                    dbc.f.get());
-                }
-              else if( dbc.f_fem )
-                {
-                  if( dbc.g_fem )
-                    new_dbc = new DirichletBoundary(dbc.b,
-                                                    vars_to_add,
-                                                    this->system(),
-                                                    dbc.f_fem.get(),
-                                                    dbc.g_fem.get());
-
-                  else
-                    new_dbc = new DirichletBoundary(dbc.b,
-                                                    vars_to_add,
-                                                    this->system(),
-                                                    dbc.f_fem.get());
-                }
-              else
-                libmesh_error_msg("Could not find valid boundary function!");
-
-              new_dbcs.push_back(new_dbc);
-            }
-        }
-
-      // Now add the new DBCs for the "dot" vars to the DofMap
-      std::vector<DirichletBoundary*>::iterator new_dbc_it =
-        new_dbcs.begin();
-
-      for( ; new_dbc_it != new_dbcs.end(); ++new_dbc_it )
-        {
-          const DirichletBoundary & dbc = *(*new_dbc_it);
-          this->system().get_dof_map().add_dirichlet_boundary(dbc);
-          delete *new_dbc_it;
-        }
-
-    } // if(all_dbcs)
 }
 
 } // namespace libMesh
